Fixed SDL_RWops for SDL 1.3 (2)

+++ src/util/bundle.c
@@ -186,7 +186,7 @@ FILE *bnd_locate(Bundle *bundle, const char *filename, int static_handle)
 
 
 #if SDL_VERSION_ATLEAST(1,3,0)
-static int bnd_seek(struct SDL_RWops *context, long offset, int whence)
+static long bnd_seek(struct SDL_RWops *context, long offset, int whence)
 #else
 static int bnd_seek(struct SDL_RWops *context, int offset, int whence)
 #endif
@@ -218,7 +218,7 @@ static int bnd_seek(struct SDL_RWops *context, int offset, int whence)
 
 
 #if SDL_VERSION_ATLEAST(1,3,0)
-static int bnd_read(struct SDL_RWops *context, void *ptr, size_t size, size_t num)
+static size_t bnd_read(struct SDL_RWops *context, void *ptr, size_t size, size_t num)
 #else
 static int bnd_read(struct SDL_RWops *context, void *ptr, int size, int num)
 #endif
